Validation

+++ resources/vue-components/content/auth/SignUpComponent.vue
@@ -32,19 +32,21 @@
                 <input v-model="password_confirmation" type="password" class="form-control"
                        id="exampleInputConfirmPassword1">
             </div>
-            <div class="mb-3">
-                <label for="exampleSelectLibrary" class="form-label">Choose your library</label>
-                <select class="form-control" id="exampleSelectLibrary">
-                    <option>1</option>
-                </select>
-            </div>
+
             <div class="mb-3"><span class="text-danger">*</span>All fields are required to be filled in</div>
             <div class="mb-3 form-check">
+
                 <input type="checkbox" class="form-check-input" id="exampleCheck1">
                 <label class="form-check-label" for="exampleCheck1">Check me out</label>
             </div>
+            <div v-if="errors.length >= 1" class="mb-3">
+                <template v-for="error in errors">
+                    <span class="text-danger">{{ error }}</span><br>
+                </template>
+            </div>
             <button @click="register" type="submit" class="btn btn-success">Sign Up</button>
         </div>
+
     </section>
 </template>
 
@@ -62,31 +64,53 @@ export default {
             password_confirmation: '',
             role_id: 1,
 
-            checkMeOut: ''
+            checkMeOut: '',
+
+            errors: []
         }
     },
 
     methods: {
         register() {
             if (this.checkMeOut != null) {
-                axios.get('/sanctum/csrf-cookie')
-                    .then(response => {
+                if ((this.email.length !==null && this.password.length >= 8) && this.password === this.password_confirmation) {
+                    axios.get('/sanctum/csrf-cookie')
+                        .then(response => {
 
-                        console.log(this.name, this.email, this.password, this.password_confirmation, this.role_id);
-                        axios.post('/register', {
-                            name: this.name,
-                            email: this.email,
-                            password: this.password,
-                            password_confirmation: this.password_confirmation,
-                            role_id: this.role_id
-                        })
-                            .then(response => {
-                                console.log('good');
-                                localStorage.setItem('x_xsrf_token', response.config.headers['X-XSRF-TOKEN'])
-                                router.push({name: 'books.index'})
+                            console.log(this.name, this.email, this.password, this.password_confirmation, this.role_id);
+                            axios.post('/register', {
+                                name: this.name,
+                                email: this.email,
+                                password: this.password,
+                                password_confirmation: this.password_confirmation,
+                                role_id: this.role_id
                             })
+                                .then(response => {
+                                    localStorage.setItem('x_xsrf_token', response.config.headers['X-XSRF-TOKEN'])
+                                    router.push({name: 'books.index'})
+                                })
+                        }).catch(response => {
+                        console.error('ERROR')
+                        this.errors.push('An incorrectly entered email address or password.')
+
                     })
+                } else this.validation()
+
+            } else this.errors.push('Confirm by clicking on the "Сheck me out" button')
+        },
+
+        validation() {
+            if (this.email.length < 1 || this.password.length < 1) {
+                this.errors.push('Fields should not be empty.')
             }
+            if (this.password.length < 8) {
+                this.errors.push('The password must be at least 8 characters long.')
+            }
+            if (this.password !== this.password_confirmation) {
+                this.errors.push('Passwords don\'t match')
+            }
+
+
         }
     }
 }
